add img-analyzer
reorganize folders and archives
