--- IaC/spark_code/staged_to_consumed/requirements.txt
@@ -1,5 +1,0 @@
-boto3
-pyspark
-pandas
-pyarrow
-fastparquet
