--- IaC/spark_code/staged_to_consumed/src/utils/utils.py
@@ -1,7 +1,0 @@
-import boto3
-
-def get_recent_file(bucket_name):
-    s3 = boto3.client('s3')
-    arquivos = s3.list_objects_v2(Bucket=bucket_name)['Contents']
-    
-    return max(arquivos, key=lambda x: x['LastModified'])['Key']
