+++ IaC/terraform/src/iot/raw_staged/raw_to_staged_payload.zip
