+++ IaC/terraform/src/iot/staged_consumed/requirements.txt
@@ -1,0 +1,1 @@
+awswrangler
