+++ IaC/terraform/src/iot/staged_consumed/staged_to_consumed.zip
