+++ IaC/terraform/src/process_img/requirements.txt
@@ -1,0 +1,2 @@
+matplotlib
+opencv-python
